bank interface

+++ class/Banker.class.php
@@ -5,9 +5,9 @@ class Banker implements Bank
 {
     private Wallet $myWallet;
 
-    private double $buffer;
+    private float $buffer;
 
-    public function storeMoney(double $amount)
+    public function storeMoney(float $amount)
     {
         $buffer += $amount;
